FFM-11022 Rename metrics file
