+++ src/metrics.ts
@@ -1,4 +1,4 @@
-import * as events from 'events';
+import * as events from "events";
 import {
   FEATURE_IDENTIFIER_ATTRIBUTE,
   FEATURE_NAME_ATTRIBUTE,
@@ -9,26 +9,23 @@ import {
   SDK_TYPE_ATTRIBUTE,
   SDK_VERSION_ATTRIBUTE,
   TARGET_ATTRIBUTE,
-  VARIATION_IDENTIFIER_ATTRIBUTE,
-} from './constants';
+  VARIATION_IDENTIFIER_ATTRIBUTE
+} from "./constants";
 import {
   Configuration,
   FeatureConfig,
+  KeyValue,
   Metrics,
   MetricsApi,
-  Variation,
-  TargetData,
-  KeyValue,
   MetricsData,
   MetricsDataMetricsTypeEnum,
-} from './openapi';
-import { Options, Target } from './types';
-import { VERSION } from './version';
-import {
-  infoMetricsSuccess,
-  infoMetricsThreadExited,
-  warnPostMetricsFailed,
-} from './sdk_codes';
+  TargetData,
+  Variation
+} from "./openapi";
+import { Options, Target } from "./types";
+import { VERSION } from "./version";
+import { infoMetricsSuccess, infoMetricsThreadExited, warnPostMetricsFailed } from "./sdk_codes";
+import { Logger } from "./log";
 
 export enum MetricEvent {
   READY = 'metrics_ready',
@@ -45,202 +42,180 @@ interface AnalyticsEvent {
 export interface MetricsProcessorInterface {
   start(): void;
   close(): void;
-  enqueue(
-    target: Target,
-    featureConfig: FeatureConfig,
-    variation: Variation,
-  ): void;
+  enqueue(target: Target, featureConfig: FeatureConfig, variation: Variation): void;
 }
 
-export const MetricsProcessor = (
-  environment: string,
-  cluster = '1',
-  conf: Configuration,
-  options: Options,
-  eventBus: events.EventEmitter,
-  closed = false,
-): MetricsProcessorInterface => {
-  const data: Map<string, AnalyticsEvent> = new Map<string, AnalyticsEvent>();
-  let syncInterval: NodeJS.Timeout;
+export class MetricsProcessor implements MetricsProcessorInterface {
+  private data: Map<string, AnalyticsEvent> = new Map();
+  private syncInterval?: NodeJS.Timeout;
+  private api: MetricsApi;
+  private readonly log: Logger;
 
-  const configuration = new Configuration({
-    ...conf,
-    basePath: options.eventsUrl,
-  });
-  const api = new MetricsApi(configuration);
-  const log = options.logger;
+  constructor(
+    private environment: string,
+    private cluster: string = "1",
+    private conf: Configuration,
+    private options: Options,
+    private eventBus: events.EventEmitter,
+    private closed: boolean = false
+  ) {
+    const configuration = new Configuration({
+      ...this.conf,
+      basePath: options.eventsUrl,
+    });
 
-  const enqueue = (
-    target: Target,
-    featureConfig: FeatureConfig,
-    variation: Variation,
-  ): void => {
-    const event: AnalyticsEvent = {
-      target,
-      featureConfig,
-      variation,
-      count: 0,
-    };
+    this.api = new MetricsApi(configuration);
+    this.log = options.logger;
+  }
 
-    const key = _formatKey(event);
-    const found = data.get(key);
+  start(): void {
+    this.log.info('Starting MetricsProcessor with request interval: ', this.options.eventsSyncInterval);
+    this.syncInterval = setInterval(() => this._send(), 10000);
+    this.eventBus.emit(MetricEvent.READY);
+  }
+
+  close(): void {
+    this.log.info('Closing MetricsProcessor');
+    if (this.syncInterval) {
+      clearInterval(this.syncInterval);
+    }
+    this._send();
+    this.closed = true;
+    this.log.info('MetricsProcessor closed');
+    infoMetricsThreadExited(this.log);
+  }
+
+  enqueue(target: Target, featureConfig: FeatureConfig, variation: Variation): void {
+    const event: AnalyticsEvent = { target, featureConfig, variation, count: 0 };
+    const key = this._formatKey(event);
+    const found = this.data.get(key);
     if (found) {
       found.count++;
     } else {
       event.count = 1;
-      data.set(key, event);
+      this.data.set(key, event);
     }
-  };
+  }
 
-  const _formatKey = (event: AnalyticsEvent): string => {
+  private _formatKey(event: AnalyticsEvent): string {
     const feature = event.featureConfig.feature;
     const variation = event.variation.identifier;
     const value = event.variation.value;
-    const target = GLOBAL_TARGET;
-    return `${feature}/${variation}/${value}/${target}`;
-  };
+    return `${feature}/${variation}/${value}/${GLOBAL_TARGET}`;
+  }
 
-  const _summarize = (): Metrics | unknown => {
-    const targetData: TargetData[] = [];
-    const metricsData: MetricsData[] = [];
+  private _summarize(): Metrics | unknown {
+      const targetData: TargetData[] = [];
+      const metricsData: MetricsData[] = [];
 
-    // clone map and clear data
-    const clonedData = new Map(data);
-    data.clear();
+      // clone map and clear data
+      const clonedData = new Map(this.data);
+      this.data.clear();
 
-    for (const event of clonedData.values()) {
-      if (event.target && !event.target.anonymous) {
-        let targetAttributes: KeyValue[] = [];
-        if (event.target.attributes) {
-          targetAttributes = Object.entries(event.target.attributes).map(
-            ([key, value]) => {
-              const stringValue =
-                value === null || value === undefined
-                  ? ''
-                  : valueToString(value);
-              return { key, value: stringValue };
-            },
-          );
+      for (const event of clonedData.values()) {
+        if (event.target && !event.target.anonymous) {
+          let targetAttributes: KeyValue[] = [];
+          if (event.target.attributes) {
+            targetAttributes = Object.entries(event.target.attributes).map(
+              ([key, value]) => {
+                const stringValue =
+                  value === null || value === undefined
+                    ? ''
+                    : this.valueToString(value);
+                return { key, value: stringValue };
+              },
+            );
+          }
+
+          let targetName = event.target.identifier;
+          if (event.target.name) {
+            targetName = event.target.name;
+          }
+
+          const td: TargetData = {
+            identifier: event.target.identifier,
+            name: targetName,
+            attributes: targetAttributes,
+          };
+          targetData.push(td);
         }
 
-        let targetName = event.target.identifier;
-        if (event.target.name) {
-          targetName = event.target.name;
-        }
+        const metricsAttributes: KeyValue[] = [
+          {
+            key: FEATURE_IDENTIFIER_ATTRIBUTE,
+            value: event.featureConfig.feature,
+          },
+          {
+            key: FEATURE_NAME_ATTRIBUTE,
+            value: event.featureConfig.feature,
+          },
+          {
+            key: VARIATION_IDENTIFIER_ATTRIBUTE,
+            value: event.variation.identifier,
+          },
+          {
+            key: SDK_TYPE_ATTRIBUTE,
+            value: SDK_TYPE,
+          },
+          {
+            key: SDK_LANGUAGE_ATTRIBUTE,
+            value: SDK_LANGUAGE,
+          },
+          {
+            key: SDK_VERSION_ATTRIBUTE,
+            value: VERSION,
+          },
+          {
+            key: TARGET_ATTRIBUTE,
+            value: event?.target?.identifier ?? null,
+          },
+        ];
 
-        const td: TargetData = {
-          identifier: event.target.identifier,
-          name: targetName,
-          attributes: targetAttributes,
+        // private target attributes
+        // need more info
+
+        const md: MetricsData = {
+          timestamp: Date.now(),
+          count: event.count,
+          metricsType: MetricsDataMetricsTypeEnum.Ffmetrics,
+          attributes: metricsAttributes,
         };
-        targetData.push(td);
+        metricsData.push(md);
       }
-
-      const metricsAttributes: KeyValue[] = [
-        {
-          key: FEATURE_IDENTIFIER_ATTRIBUTE,
-          value: event.featureConfig.feature,
-        },
-        {
-          key: FEATURE_NAME_ATTRIBUTE,
-          value: event.featureConfig.feature,
-        },
-        {
-          key: VARIATION_IDENTIFIER_ATTRIBUTE,
-          value: event.variation.identifier,
-        },
-        {
-          key: SDK_TYPE_ATTRIBUTE,
-          value: SDK_TYPE,
-        },
-        {
-          key: SDK_LANGUAGE_ATTRIBUTE,
-          value: SDK_LANGUAGE,
-        },
-        {
-          key: SDK_VERSION_ATTRIBUTE,
-          value: VERSION,
-        },
-        {
-          key: TARGET_ATTRIBUTE,
-          value: event?.target?.identifier ?? null,
-        },
-      ];
-
-      // private target attributes
-      // need more info
-
-      const md: MetricsData = {
-        timestamp: Date.now(),
-        count: event.count,
-        metricsType: MetricsDataMetricsTypeEnum.Ffmetrics,
-        attributes: metricsAttributes,
+      return {
+        targetData: targetData,
+        metricsData: metricsData,
       };
-      metricsData.push(md);
-    }
-    return {
-      targetData: targetData,
-      metricsData: metricsData,
-    };
-  };
+  }
 
-  const _send = (): void => {
-    if (closed) {
-      log.debug('SDK has been closed before metrics can be sent');
+  private async _send(): Promise<void> {
+    if (this.closed) {
+      this.log.debug('SDK has been closed before metrics can be sent');
       return;
     }
 
-    if (data.size == 0) {
-      log.debug('No metrics to send in this interval');
+    if (this.data.size === 0) {
+      this.log.debug('No metrics to send in this interval');
       return;
     }
 
-    const metrics: Metrics = _summarize();
+    const metrics: Metrics = this._summarize();
 
-    log.debug('Start sending metrics data');
-    api
-      .postMetrics(environment, cluster, metrics)
-      .then((response) => {
-        log.debug('Metrics server returns: ', response.status);
-        infoMetricsSuccess(log);
-        if (response.status >= 400) {
-          log.error(
-            'Error while sending metrics data with status code: ',
-            response.status,
-          );
-        }
-      })
-      .catch((error: Error) => {
-        warnPostMetricsFailed(`${error}`, log);
-        log.debug('Metrics server returns error: ', error);
-      });
-  };
+    this.log.debug('Start sending metrics data');
+    try {
+      const response = await this.api.postMetrics(this.environment, this.cluster, metrics);
+      this.log.debug('Metrics server returns: ', response.status);
+      infoMetricsSuccess(this.log);
+      if (response.status >= 400) {
+        this.log.error('Error while sending metrics data with status code: ', response.status);
+      }
+    } catch (error) {
+      warnPostMetricsFailed(`${error}`, this.log);
+      this.log.debug('Metrics server returns error: ', error);
+    }
+  }
 
-  const start = (): void => {
-    log.info(
-      'Starting MetricsProcessor with request interval: ',
-      options.eventsSyncInterval,
-    );
-    syncInterval = setInterval(_send, options.eventsSyncInterval);
-    eventBus.emit(MetricEvent.READY);
-  };
-
-  const valueToString = (value: any): string => {
-    return typeof value === 'object' && !Array.isArray(value) ? JSON.stringify(value) : String(value)
-  };
-
-  const close = (): void => {
-    log.info('Closing MetricsProcessor');
-    clearInterval(syncInterval);
-    _send();
-    closed = true;
-    log.info('MetricsProcessor closed');
-    infoMetricsThreadExited(log);
-  };
-
-  return {
-    start,
-    close,
-    enqueue,
-  };
-};
+  private valueToString(value: any): string {
+    return typeof value === 'object' && !Array.isArray(value) ? JSON.stringify(value) : String(value);
+  }
+}
